Three-axis accelerometer simulation, rotating to and from rocket frame.
Besides handling the difference between the rocket and Earth frames of
reference, the BPF now validates all three axes of the acceleration
estimate. It uses a distinct measurement variance estimate for each
individual accelerometer on the rocket.

We still don't change the rocket's orientation at any time during the
simulation, nor add noise to anything but the Z acceleration estimate in
the BPF. But if we did, this code should work correctly.