Bump accelerometer measurements away from gravity by 1g.
There is a truly marvelous proof that this is correct, which this commit
message is too small to contain.

Hint: Gravity affects the test mass in an accelerometer and the rocket
frame equally, while other forces affect the rocket frame and then have to
be transmitted to the accelerometer from there.

Hint: On the ground, the accelerometer reads 1g upward force due to the
ground holding the rocket up. If the rocket were hovering, with enough
thrust to hold it still, the acceleration due to thrust would be 1g upward
to cancel the 1g downward of gravity, and the accelerometer would read a
1g upward force there too. If the rocket were in free-fall (and not
affected by drag), then the rocket frame and accelerometer are
accelerating equally and the accelerometer would read 0.