Add copyright notice and license header to all source files
Bart says that we can mark "Portland State Aerospace Society" as the
copyright holder, which nicely avoids the problem of updating the notice
every time a new person modifies a file.  Also, he says we don't need to
bother updating the date every year, either.

For the files which have a copyright notice from Bart, leave his name in
as well, at his request, but modify the license notice to match the
rest.

Reviewed-by: Jamey Sharp <jamey@minilop.net>