Pre-multiply semicircles into radians.
Radians are the standard unit; semicircles are just weird. Store the
sensible unit in "struct ephemeris".

Writing this patch led me to notice the bug fixed in the previous
commit, which I take as evidence that we want to get to radians as
quickly as possible.

As a bonus this moves a handful of constant multiplies from the
frequently executed gps_satellite_position to the infrequently
executed parse_ephemeris.