Separate vec3 struct from the union that allows array access.
We almost always access our vectors by named component. In those rare
cases where we want to loop over the components by index, we can
construct a temporary (union vec_array) and get the array form that way.

Syntactically, this change is convenient because it means we can use
single braces around vec3 initializers, instead of the double braces
required when it was a struct nested inside a union.

However the real reason I made this change is I noticed GCC was
optimizing poorly through vector temporaries. Based on its output, I'm
guessing it thought the presence of the array member meant every vec3
had to have a contiguous representation in memory, because it allocated
stack space for every intermediate vec3 temporary and memcpy'd between
them frequently.

After this change GCC is willing to keep vector components in registers
and avoids moving values unnecessarily, for sometimes huge wins in
optimization.